+++ localstack/utils/testing/snapshots/__init__.py
@@ -1,0 +1,3 @@
+from .prototype import SnapshotAssertionError, SnapshotMatchResult, SnapshotSession
+
+__all__ = [SnapshotSession, SnapshotMatchResult, SnapshotAssertionError]
