+++ localstack/utils/testing/snapshots/prototype.py
